Take logging code out of kernel, use LED
I can't call my Logger functions in the kernel code itself because
interrupts are disabled and it can lock up since it's waiting for the
lock to be released. Will have to work on that later.
Instead I'm turning on and off an LED on pin 11.
It's not really working yet, apparently setting SP messes it up, even if
I don't change it (to my knowledge)!